refactor(introduce): 파일 명 변경, 공통 사용 style  common 파일 생성
