+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
-import styled from '@emotion/styled';
 import type { NextPage } from 'next';
 
+import { Container } from '@/styles/common';
 import Introduce from 'templates/main/Introduce/Introduce';
 
 const Home: NextPage = () => {
@@ -12,12 +12,3 @@ const Home: NextPage = () => {
 };
 
 export default Home;
-
-const Container = styled('div')({
-  display: 'flex',
-  flexDirection: 'column',
-  width: '100%',
-  justifyContent: 'center',
-  alignItems: 'center',
-  gap: '4px',
-});
